requirements.txt
sounddevice==0.5.1
keyboard==0.13.5
numpy==2.1.3  
pydub==0.25.1
psutil==7.0.0
mutagen==1.47.0
pillow==11.1.0
pystray==0.19.5